OP-1216 Convert FLOAT/DOUBLE money and quantity columns to DECIMAL/INT
Convert monetary and exact-quantity fields from floating point to
java.math.BigDecimal (DECIMAL(19,2)), and pure unit-count fields to int,
to guarantee exact precision for financial and stock values.

- Models -> BigDecimal: Price.price, Bill.amount/balance, BillItems,
  BillPayments, MovementWard.quantity, MedicalWard.out_quantity, Lot.cost.
- Models -> int: Medical in/out/min qty, MedicalWard.in_quantity.
- Managers/services: arithmetic reworked on BigDecimal (compareTo for
  zero checks, setScale(2, HALF_UP) preserved, price-step rounding kept).
  Ward stock repository out-quantity queries now take BigDecimal (no
  double round-trip); in-quantity queries take int.
- SQL: new step_a120_decimal_precision.sql, registered in step_04.
- Tests updated for the new types.

+++ src/main/java/org/isf/medicalinventory/manager/MedicalInventoryManager.java
@@ -21,6 +21,7 @@
  */
 package org.isf.medicalinventory.manager;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -643,7 +644,7 @@ public class MedicalInventoryManager {
 			Lot currentLot = medicalInventoryRow.getLot();
 			if (movQuantity != 0) {
 				insertedMovements
-					.add(movWardBrowserManager.newMovementWard(new MovementWard(selectedWard, now, false, null, 0, 0, reason, medical, movQuantity,
+					.add(movWardBrowserManager.newMovementWard(new MovementWard(selectedWard, now, false, null, 0, 0, reason, medical, BigDecimal.valueOf(movQuantity),
 						MessageBundle.getMessage("angal.medicalstockward.rectify.pieces"), currentLot)));
 			}
 		}
